FIFA-WorldCup-2026/système-de-reconnaissance-hybride/face-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ML/IA
tensorflow==2.15.0
opencv-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
